created pill-container component from listbox-of-pill-options
