+++ components/pill-container/__docs__/site-stories.js
@@ -5,7 +5,9 @@
 
 const siteStories = [
 	require('raw-loader!@salesforce/design-system-react/components/pill/__examples__/base.jsx'),
-	require('raw-loader!@salesforce/design-system-react/components/pill/__examples__/icon.jsx'),
+	require('raw-loader!@salesforce/design-system-react/components/pill/__examples__/icons.jsx'),
+	require('raw-loader!@salesforce/design-system-react/components/pill/__examples__/avatars.jsx'),
+	require('raw-loader!@salesforce/design-system-react/components/pill/__examples__/bare.jsx'),
 ];
 
 module.exports = siteStories;
