+++ components/pill-container/__examples__/avatars.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import ListboxOfPillOptions from '~/components/pill/listbox-of-pill-options';
 import IconSettings from '~/components/icon-settings';
+import PillContainer from '~/components/pill-container';
 
 class Example extends React.Component {
-	static displayName = 'BarePillListboxExample';
+	static displayName = 'PillWithAvatarListboxExample';
 
 	constructor(props) {
 		super(props);
@@ -12,14 +12,22 @@ class Example extends React.Component {
 		this.state = {
 			options: [
 				{
+					avatar: {
+						imgSrc:
+							'https://lightningdesignsystem.com/assets/images/avatar1.jpg',
+						title: 'User 1',
+					},
 					id: '1',
-					isBare: true,
 					label: 'Pill Label 1',
 					title: 'Full pill label verbiage mirrored here',
 				},
 				{
+					avatar: {
+						imgSrc:
+							'https://lightningdesignsystem.com/assets/images/avatar2.jpg',
+						title: 'User 2',
+					},
 					id: '2',
-					isBare: true,
 					label: 'Pill Label 2',
 					title: 'Full pill label verbiage mirrored here',
 				},
@@ -51,7 +59,7 @@ class Example extends React.Component {
 						<h3 className="slds-text-heading_small">Static Examples</h3>
 					</div>
 					<div className="slds-grid slds-grid_vertical-align-start">
-						<ListboxOfPillOptions
+						<PillContainer
 							options={this.state.options}
 							onClickPill={this.onClickPill}
 							onRemovePill={this.onRemovePill}
