+++ components/pill-container/__examples__/bare.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import ListboxOfPillOptions from '~/components/pill/listbox-of-pill-options';
 import IconSettings from '~/components/icon-settings';
+import PillContainer from '~/components/pill-container';
 
 class Example extends React.Component {
-	static displayName = 'PillListboxExample';
+	static displayName = 'BarePillListboxExample';
 
 	constructor(props) {
 		super(props);
@@ -13,11 +13,13 @@ class Example extends React.Component {
 			options: [
 				{
 					id: '1',
+					isBare: true,
 					label: 'Pill Label 1',
 					title: 'Full pill label verbiage mirrored here',
 				},
 				{
 					id: '2',
+					isBare: true,
 					label: 'Pill Label 2',
 					title: 'Full pill label verbiage mirrored here',
 				},
@@ -49,7 +51,7 @@ class Example extends React.Component {
 						<h3 className="slds-text-heading_small">Static Examples</h3>
 					</div>
 					<div className="slds-grid slds-grid_vertical-align-start">
-						<ListboxOfPillOptions
+						<PillContainer
 							options={this.state.options}
 							onClickPill={this.onClickPill}
 							onRemovePill={this.onRemovePill}
