+++ components/pill-container/docs.json
@@ -1,0 +1,7 @@
+{
+    "component": "pill-listbox",
+    "status": "prototype",
+    "display-name": "Pill Listbox",
+    "SLDS-component-path": "/components/pill-listbox",
+    "url-slug": "pill-listbox"
+}
