+++ components/pill/__tests__/pill.browser-test.jsx
@@ -1,17 +1,12 @@
 /* eslint-disable react/no-find-dom-node */
 import React from 'react';
-import TestUtils from 'react-dom/test-utils';
-import IconSettings from '~/components/icon-settings';
 import SLDSPill from '~/components/pill';
 import SLDSIcon from '~/components/icon';
 import SLDSAvatar from '~/components/avatar';
-import SLDSListboxOfPillOptions from '~/components/pill/listbox-of-pill-options';
 import { mountComponent, unmountComponent } from '~/tests/enzyme-helpers';
 import { expect } from 'chai';
 
-const { Simulate } = TestUtils;
-
-describe.only('SLDSPill', () => {
+describe('SLDSPill', () => {
 	const LABEL = 'Pill Label';
 	const LABEL_TITLE = 'Full pill label verbiage mirrored here';
 
@@ -305,86 +300,6 @@ describe.only('SLDSPill', () => {
 			expect(removeButton).to.exist;
 			removeButton.simulate('click');
 			expect(onRemove.calledOnce).to.be.true;
-		});
-	});
-
-	describe.only('Listbox of Pill Options', () => {
-		let clickData = {};
-		let removeData = {};
-
-		beforeEach(
-			mountComponent(
-				<IconSettings iconPath="/assets/icons">
-					<SLDSListboxOfPillOptions
-						options={[
-							{
-								id: '1',
-								label: 'Pill Label 1',
-								title: 'Full pill label verbiage mirrored here',
-							},
-							{
-								icon: {
-									category: 'standard',
-									name: 'account',
-								},
-								id: '2',
-								label: 'Pill Label 2',
-								title: 'Full pill label verbiage mirrored here',
-							},
-							{
-								avatar: {
-									imgSrc:
-										'https://lightningdesignsystem.com/assets/images/avatar1.jpg',
-									title: 'User 3',
-								},
-								id: '3',
-								label: 'Pill Label 3',
-								title: 'Full pill label verbiage mirrored here',
-							},
-						]}
-						onClickPill={(event, data) => {
-							clickData = data;
-						}}
-						onRemovePill={(event, data) => {
-							removeData = data;
-						}}
-					/>
-				</IconSettings>
-			)
-		);
-
-		afterEach(unmountComponent);
-
-		it('Renders the Listbox of Pill Options correctly', function() {
-			let currentPill = 1;
-
-			expect(this.wrapper.find('.slds-pill_container').length).to.eql(1);
-			expect(this.wrapper.find('ul.slds-listbox').length).to.eql(1);
-
-			this.wrapper.find('ul.slds-listbox .slds-pill').forEach((pill) => {
-				expect(pill.find('.slds-pill__label').text()).to.eql(
-					`Pill Label ${currentPill}`
-				);
-
-				Simulate.click(pill.getDOMNode());
-				expect(clickData.option.id).to.eql(`${currentPill}`);
-
-				Simulate.click(pill.find('.slds-pill__remove').getDOMNode());
-				expect(removeData.option.id).to.eql(`${currentPill}`);
-
-				if (currentPill === 2) {
-					expect(
-						pill.find('.slds-icon_container.slds-icon-standard-account').length
-					).to.eql(1);
-				} else if (currentPill === 3) {
-					expect(
-						pill.find('.slds-avatar.slds-avatar_circle.slds-avatar_medium')
-							.length
-					).to.eql(1);
-				}
-
-				currentPill++;
-			});
 		});
 	});
 });
