+++ components/site-stories.js
@@ -38,6 +38,7 @@ const documentationSiteLiveExamples = {
 	'page-header': require('@salesforce/design-system-react/components/page-header/__docs__/site-stories.js'),
 	panel: require('@salesforce/design-system-react/components/panel/__docs__/site-stories.js'),
 	pill: require('@salesforce/design-system-react/components/pill/__docs__/site-stories.js'),
+	'pill-container': require('@salesforce/design-system-react/components/pill-container/__docs__/site-stories.js'),
 	popover: require('@salesforce/design-system-react/components/popover/__docs__/site-stories.js'),
 	'progress-indicator': require('@salesforce/design-system-react/components/progress-indicator/__docs__/site-stories.js'),
 	'progress-ring': require('@salesforce/design-system-react/components/progress-ring/__docs__/site-stories.js'),
