+++ components/storybook-stories.js
@@ -43,6 +43,7 @@ export Notification from '../components/notification/__docs__/storybook-stories'
 export PageHeader from '../components/page-header/__docs__/storybook-stories';
 export Panel from '../components/panel/__docs__/storybook-stories';
 export Pill from '../components/pill/__docs__/storybook-stories';
+export PillContainer from '../components/pill-container/__docs__/storybook-stories';
 export Popover from '../components/popover/__docs__/storybook-stories';
 export ProgressIndicator from '../components/progress-indicator/__docs__/storybook-stories';
 export ProgressRing from '../components/progress-ring/__docs__/storybook-stories';
