+++ utilities/constants.js
@@ -76,6 +76,7 @@ export const PANEL_FILTER_LIST = 'SLDSFilterList';
 export const PANEL_FILTER_LIST_HEADING = 'SLDSPanelFilterListHeading';
 export const PAGE_HEADER = 'SLDSPageHeader';
 export const PILL = 'SLDSPill';
+export const PILL_CONTAINER = 'SLDSPillContainer';
 export const POPOVER = 'SLDSPopover';
 export const POPOVER_TOOLTIP = 'SLDSPopoverTooltip';
 export const PROGRESS_INDICATOR = 'SLDSProgressIndicator';
